test for stale and delay

+++ src/backend/apps/webcam/tasks.py
@@ -139,15 +139,6 @@ def update_all_webcam_data():
             raise CancelExecution()
 
         current_time = datetime.datetime.now(tz=ZoneInfo("America/Vancouver"))
-        # if camera.https_cam:
-        #     update_cam_from_sql_db(camera.id, current_time)
-        # else:
-        #     if camera.should_update(current_time):
-        #         updated = update_single_webcam_data(camera)
-        #         if updated:
-        #             update_camera_group_id(camera)
-        
-        # bruce test
         update_cam_from_sql_db(camera.id, current_time)
         if camera.should_update(current_time):
             updated = update_single_webcam_data(camera)
@@ -512,9 +503,6 @@ def update_cam_from_sql_db(id: int, current_time: datetime.datetime):
             # Query from Cams_Live
             result_live = connection.execute(cams_live_sql, {"id": id})
             live_rows = {row.id: dict(row._mapping) for row in result_live}
-            # bruce test
-            if id == 36:
-                print(f"bruce test SQL DB returned {len(live_rows)} rows for cam ID {id}")
             update_webcam_db(id, live_rows.get(id, {}))
             return live_rows
 
@@ -527,12 +515,7 @@ def update_webcam_db(cam_id: int, cam_data: dict):
     if not timestamp_utc:
         return
 
-    # bruce test
-    if cam_id == 36:
-        print(f"bruce test update_webcam_db for cam ID {cam_id} with timestamp_utc: {timestamp_utc}")
     camera_status = calculate_camera_status(timestamp_utc)
-    if cam_id == 36:
-        print(f"bruce test camera_status: {camera_status}")
     ts_seconds = int(camera_status["timestamp"])
     dt_utc = datetime.datetime.fromtimestamp(ts_seconds, tz=ZoneInfo("UTC"))
 
@@ -623,6 +606,8 @@ def purge_old_pvc_s3_images(age: str = "24", is_pvc: bool = True):
         logger.info(f"Deleting {len(files_to_delete)} old PVC images...")
         for file_path in files_to_delete:
             try:
+                if not file_path:
+                    continue
                 os.remove(file_path)
                 logger.info(f"Deleted file from PVC: {file_path}")
             except FileNotFoundError:
